+++ lib/public/Model.js
@@ -13,7 +13,7 @@
 
 import { Loader, Observable, QueryRouter, sessionService } from '/js/src/index.js';
 
-import Logs from './views/Logs/Logs.js';
+import Logs from './views/Logs/Overview/LogsOverviewModel.js';
 import Tags from './views/Tags/Tags.js';
 import { RunsModel } from './views/Runs/RunsModel.js';
 import Subsystems from './views/Subsystems/Subsystems.js';
