+++ lib/public/views/Logs/Create/LogCreationModel.js
@@ -32,7 +32,7 @@ export class LogCreationModel extends Observable {
      * @param {string[]} preFilledRelations.lhcFillNumbers the fill numbers to link to the log being created
      * @param {string[]} preFilledRelations.environmentIds the environment ids to link to the log being created
      */
-    constructor(onCreation, parentLogId, { runNumbers, lhcFillNumbers, environmentIds }) {
+    constructor(onCreation) {
         super();
 
         this._onCreation = onCreation;
@@ -42,33 +42,43 @@ export class LogCreationModel extends Observable {
         this._creationTagsPickerModel = new TagPickerModel();
         this._creationTagsPickerModel.bubbleTo(this);
         this._creationTagsPickerModel.visualChange$.bubbleTo(this);
-
-        this._runNumbers = runNumbers.join(DATA_DISPLAY_DELIMITER);
-        this._lhcFills = lhcFillNumbers.join(DATA_DISPLAY_DELIMITER);
-        this._environments = environmentIds.join(DATA_DISPLAY_DELIMITER);
-        this._parentLogId = parentLogId;
-        if (parentLogId) {
-            this._parentLog = RemoteData.loading();
-            getRemoteData(`/api/logs/${parentLogId}`).then(
-                ({ data: parentLog }) => {
-                    this.parentLog = RemoteData.success(parentLog);
-                },
-                (errors) => {
-                    this.parentLog = RemoteData.failure(errors);
-                },
-            );
-        } else {
-            this._parentLog = RemoteData.success(null);
-        }
+        this._parentLogId = null;
 
         this._isParentLogCollapsed = true;
 
         this._attachments = [];
 
         this._createdLog = RemoteData.notAsked();
-        this._isRunNumbersReadonly = this._runNumbers;
+        this._isRunNumbersReadonly = false;
 
         this._textEditor = null;
+    }
+
+    /**
+     * Provide initial data fro creation
+     * @param {Object} [data] data
+     * @param {number} [data.parentLogId] id of parent log
+     * @param {number[]} [data.runNumbers] run numbers associated with log being created
+     * @param {number[]} [data.lhcFillNumbers] lhc fill numbers associated with log being created
+     * @param {number[]} [data.environmentIds] environment ids associated with log being created
+     * @return {void}
+     */
+    setLogCreationInitialData({ parentLogId, runNumbers, lhcFillNumbers, environmentIds }) {
+        this._runNumbers = runNumbers.join(DATA_DISPLAY_DELIMITER);
+        this._lhcFills = lhcFillNumbers.join(DATA_DISPLAY_DELIMITER);
+        this._environments = environmentIds.join(DATA_DISPLAY_DELIMITER);
+        this._isRunNumbersReadonly = Boolean(this._runNumbers);
+
+        this._parentLogId = parentLogId;
+        this._parentLog = RemoteData.loading();
+        getRemoteData(`/api/logs/${parentLogId}`).then(
+            ({ data: parentLog }) => {
+                this.parentLog = RemoteData.success(parentLog);
+            },
+            (errors) => {
+                this.parentLog = RemoteData.failure(errors);
+            },
+        );
     }
 
     /**
