+++ lib/public/views/Logs/Overview/LogsOverviewModel.js
@@ -12,21 +12,19 @@
  */
 
 import { Observable, RemoteData } from '/js/src/index.js';
-import { TagFilterModel } from '../../components/Filters/common/TagFilterModel.js';
-import { SortModel } from '../../components/common/table/SortModel.js';
-import { debounce } from '../../utilities/debounce.js';
-import { FilterInputModel } from '../../components/Filters/common/filters/FilterInputModel.js';
-import { LogCreationModel } from './Create/LogCreationModel.js';
-import { PaginationModel } from '../../components/Pagination/PaginationModel.js';
-import { getRemoteDataSlice } from '../../utilities/fetch/getRemoteDataSlice.js';
-import { LogTreeViewModel } from './Details/LogTreeViewModel.js';
+import { TagFilterModel } from '../../../components/Filters/common/TagFilterModel.js';
+import { SortModel } from '../../../components/common/table/SortModel.js';
+import { debounce } from '../../../utilities/debounce.js';
+import { FilterInputModel } from '../../../components/Filters/common/filters/FilterInputModel.js';
+import { PaginationModel } from '../../../components/Pagination/PaginationModel.js';
+import { getRemoteDataSlice } from '../../../utilities/fetch/getRemoteDataSlice.js';
 
 /**
  * Model representing handlers for log entries page
  *
  * @implements {OverviewModel}
  */
-export default class LogModel extends Observable {
+export default class LogsOverviewModel extends Observable {
     /**
      * The constructor of the Overview model object
      *
@@ -40,11 +38,6 @@ export default class LogModel extends Observable {
         this._listingTagsFilterModel = new TagFilterModel();
         this._listingTagsFilterModel.observe(() => this._applyFilters());
         this._listingTagsFilterModel.visualChange$.bubbleTo(this);
-
-        this._creationModel = null;
-
-        this._treeViewModel = new LogTreeViewModel();
-        this._treeViewModel.bubbleTo(this);
 
         this._overviewSortModel = new SortModel();
         this._overviewSortModel.observe(() => this._applyFilters(true));
@@ -72,50 +65,6 @@ export default class LogModel extends Observable {
         const updateDebounceTime = () => this._debouncedFetchAllLogs = debounce(this.fetchAllLogs.bind(this), model.inputDebounceTime);
         model.appConfiguration$.observe(() => updateDebounceTime());
         updateDebounceTime();
-    }
-
-    /**
-     * Parse the current route parameters to load the log creation model
-     *
-     * @param {object} queryParams the current query parameters
-     * @param {string} [queryParams.parentLogId] the id of log that should be the parent of the log being created
-     * @param {string|string[]} [queryParams.runNumbers] the run numbers to link to the log being created
-     * @param {string|string[]} [queryParams.lhcFillNumbers] the lhc fill numbers to link to the log being created
-     * @param {string|string[]} [queryParams.environmentIds] the environment ids to link to the log being created
-     * @return {void}
-     */
-    loadCreation(queryParams) {
-        const parentLogId = parseInt(queryParams.parentLogId, 10);
-        let { runNumbers, lhcFillNumbers, environmentIds } = queryParams;
-
-        if (!Array.isArray(runNumbers)) {
-            runNumbers = runNumbers !== undefined ? [runNumbers] : [];
-        }
-
-        if (!Array.isArray(lhcFillNumbers)) {
-            lhcFillNumbers = lhcFillNumbers !== undefined ? [lhcFillNumbers] : [];
-        }
-
-        if (!Array.isArray(environmentIds)) {
-            environmentIds = environmentIds !== undefined ? [environmentIds] : [];
-        }
-
-        this._creationModel = new LogCreationModel(
-            (logId) => this.model.router.go(`/?page=log-detail&id=${logId}`),
-            parentLogId,
-            { runNumbers, lhcFillNumbers, environmentIds },
-        );
-        this._creationModel.bubbleTo(this);
-    }
-
-    /**
-     * Load the view of the tree of the given log
-     *
-     * @param {number} selectedLogId the id of the log for which tree must be displayed
-     * @return {void}
-     */
-    loadTreeView(selectedLogId) {
-        this._treeViewModel.selectedLogId = selectedLogId;
     }
 
     /**
@@ -152,14 +101,6 @@ export default class LogModel extends Observable {
         }
 
         this.notify();
-    }
-
-    /**
-     * Getter for all the data
-     * @returns {RemoteData} Returns all of the filtered logs
-     */
-    getLogs() {
-        return this._logs;
     }
 
     /**
@@ -381,24 +322,6 @@ export default class LogModel extends Observable {
      */
     get contentFilter() {
         return this._contentFilter;
-    }
-
-    /**
-     * Return the log creation model
-     *
-     * @return {LogCreationModel|null} the model
-     */
-    get creationModel() {
-        return this._creationModel;
-    }
-
-    /**
-     * Returns the tree view model
-     *
-     * @return {LogTreeViewModel} the model
-     */
-    get treeViewModel() {
-        return this._treeViewModel;
     }
 
     /**
